requirements.txt
# requirements.txt
langchain-community>=0.0.10
langchain>=0.1.0
openai>=1.3.0
python-dotenv>=1.0.0
streamlit>=1.29.0
flask>=3.0.0
python-decouple>=3.8
sqlalchemy>=2.0.23
google-generativeai==0.3.0
groq                        
